I wrote a basic font end that displays tweets from users you follow. It makes calls to the server to check if any new tweets have been posted by users you follow. If they have a button to load them is shown.
The server persists all tweets by users and runs the tweet generation
function for other users. It is run through ngronk using 'npm run dev'
to initiate the server. When initialized the server generates a few
tweets to supply the app. Also when the server is initiated you need
to copy the URL to the app in app/app.js at the top. If this isnt done
the app WONT WORK. So that is crucial.